feature: implement user profile and other features
implement view function for reaturning the user profile. implement view function for returning the view for editing user profiles. add view function for displaying the last_seen for each user such that immediately after a user sends a request to the server we store that time as the last seen for the said user. Users can only see the option to edit their profile if they are logged in and viewing their own profile.

[#chapter6]